LayoutTests/http/tests/fileapi/create-blob-url-from-data-url-expected.txt

Test case for checking blob URL not allowed to be created from the code running from data URI

PASS: no blob URL is created
